Previous Peak Detector was making the compressor to diminish the whole signal
The preivous process() function of the Peak Detector class was only updating the singal every n samples, so everytime the signal goes above the threshold (e.g 0.8) it triggers the compressor but for the next n samples the signal was still returning the same 0.8 even if it was quieter.

Also fast transients were missed if they don't line up with the update window.